GatewayPaymentDetails component was extracted
